Improved Command filter
Added options and flags. Options will consume both keyword and value right next to, flags will just consume the keyword itself.

`message.command` is now a dict : the "old" list can be found in `message.command["raw"]`. Options, if found, will be inserted in the dict with the name they were registered with. Flags will be added to a list in `message.command["flags"]`. The remaining command can be accessed as a list in `message.command["cmd"]` or as one joined string in `message.command["arg"]`. Note that neither of these will be present if options and flags will consume the command list completely.